# public/id/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NirKyy API",
    "version": "1.0.0",
    "description": "Dokumentasi API untuk Pengembang",
    "contact": {
      "name": "NirKyy",
      "url": "https://wa.me/6283894391287"
    }
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "tags": [
    {
      "name": "Utama",
      "description": "Endpoint utama"
    },
    {
      "name": "Spotify",
      "description": "Endpoint terkait Spotify"
    }
  ],
  "paths": {
    "/main/ping": {
      "get": {
        "tags": ["Utama"],
        "summary": "Cek Ping API",
        "description": "Mengembalikan status 'ok' jika API berjalan.",
        "responses": {
          "200": {
            "description": "Respon sukses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "message": {
                      "type": "string",
                      "example": "pong"
                    },
                    "author": {
                      "type": "string",
                      "example": "NirKyy"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/spotify/track": {
      "get": {
        "tags": ["Spotify"],
        "summary": "Unduh lagu Spotify",
        "description": "Mengunduh lagu dari Spotify dan mengembalikan tautan unduhan langsung.",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "description": "URL lagu Spotify yang akan diunduh.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Respon sukses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "artist": {
                      "type": "string"
                    },
                    "duration": {
                      "type": "string"
                    },
                    "image": {
                      "type": "string"
                    },
                    "spotifyUrl": {
                      "type": "string"
                    },
                    "downloadUrl": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "URL lagu tidak ditemukan"
          },
          "500": {
            "description": "Gagal mengunduh lagu"
          }
        }
      }
    }
  }
}